Fixed the immediate bug mentioned in the JIRA 3204 i.e. Grid needs to resize when you resize the window.
but also noticed that there was some unequal spacing and aligment issues in the top of the bar, mainly due to hardcoded width sises playing havoc in the table layout.  I simplified these and added some padding to the help icon when they appear and also made it look better when the screen is stretched.

Because of the ability to now stretch the screen width I also made the buttons take up the central space rather than justified right.