JIRA-3972 - Pivots weren't generating any data. Firstly had to jig the order in which the data is generated so the dat page is done first. Secondly you need to ensure that the columns are correctly formatted as numbers or else the sum function won't work.
Also I've tidied up with a few try...catches and removed some old commented out code.

In this changeset the only fix is that the data should now be correct and display the same as the data manager.  I have not marked the jira as fixed as the following items are still outstanding.
1) Formating not applied to the pivot table
2) Data page still visible (not sure if this is a good or a bad thing)
3) Download needs marking as safe or else the pivot doesn't display until you've said enable editing (there may also be a way of updating the pivot before streamiung teh document)