JIRA-3830 - The current table was not being remembered as you edited defintions.
It now remembers the current table ID, but also does not get reset when moving between picklists, calcs and filters. At login it defaults mto personnel records but after that will only ever be the last table you selected.

Also made the dropdown table selection immediate rather than having to click the go button.

And finally swapped the bit that gets the table list to read from the local caches version to avoid an unecessary roundtrip to the database. Also removed this stored proc as it now redundant