JIRA-3972 - Pivot tables displaying in screwy format.
There is still an outstanding issue when the user modifies the pivot in that the formatting is zapped. (These would appear to be related bugs)
http://social.technet.microsoft.com/Forums/office/en-US/d4a13831-b77b-4db5-97bd-af0e00683c8d/excel-2013-pivot-chart-formatting-lost?forum=excel
http://www.aspose.com/community/forums/thread/497759/text-wrap-setting-lost-on-pivot-refresh-in-values-columns.aspx

I think the solution is to add a pivottable style and use that instead, but that could take a little while, and I'm tired, hungry and want to go home.