JIRA-3724 - A whopping report with loads of page breaks meant the sheet naming was blowing up. Despite the check I've put in after a couple of hundred (yes this report has HUNDREDS) the name check fails, so I've put another catch all check of 100 and it just leaves the standard sheet name.
There was also a bug (unnoticed by QA) that the last page of data was not getting outputted, this was mostly to do with the conversion from vb6 where arrays are 1 bound instead of zero.

P.S. There are some outsnading issues with extra columns being displayed - these are raised as seperate bugs.
P.P.S The report in the JIRA is never going to happen in real life!