JIRA-4021 - We were getting deadlocking issues as part of a load test when releasing v8. I noted that the ones that seemed to creep up the most were the ones on tbsys_intransactiontrigger, so I've swapped this out to be a temp table which will be scoped to this transaction only.
I wanted to use the session context info which would avoid the use of a table entirely (be cleaner, safer, and more performant) however this is the mechanism that Fusion uses and so without rework there I cannot safely use. (Its annoying actually as Fusion doesn't really do anything with this apart from spoil all other uses of it. Its a bit like being back at school and the annoying bully in the corner who steals all the crayons in art class but doesn't actually use them, and so the only way I can get that nice shade of red that I'm after is to slit my own wrists!)

Anyhow. This is not a complete fix, as we seem to have deadlock issues on the main upadte statement in U02. SOme of the udfs make calls to the udftab functions and these seem to implement a table lock. In other words if a large absence update is happening any reads on absence table during that process are vunerable to a deadlock even if you are looking at a different subset of data. Two solutions to this, re-imagine how we built the udf in the triggers (this would be nice as we could tackle performance at the same time - but is likely to be a HUGE job) Secondly, Investigate the possibility of converting the udftabs into views which handle locking differently. Again, possibly a big job.

So in summary, this changeset should make this less deadlocky, but may (or may not for that matter) have an affect on performance)

If you start running some global updates on child tables (salary and absence being the biggest offenders) then you are still liable to get some deadlocks.