JIRA-3675/4055 - The group with next was not working properly, particularily if there wre two of them. Basically we now generate the output data table with data in the same columns as the output, rather than fettle which groupby and which summary when it comes to display time. This should mean output and preview now look at the same data.
I've tried to simplify the problem of having some loops starting at zero and some at 1. This is part due to having upgraded some of the components from VB6 which behaves differently to .NET.

Also included in this is a tweak to linksmain which I found was refreshing the main page for no apparent reason when hitting a report/screen from the dashboard. (I think this weas from the old msystem that needed to preserve session info - which has now been superseded by Roberto's absolutely superb mouseclicky code)

There are still complications with the output using a combination of arrays and datatables, which I'd like to clear up if I get time.

Note: This does not cover jira-3963 (excel output merging into one string)