JIRA-4121 - Crashed when drilling down into cell breakdown for absence breakdown and cross tab. Was caused by 7817 clearing up the temporary tables.
We now pull all the data into a local temp table and process that (doesn't need to re-reade the sql server 3 times!). Added a new new stored proc which does it all in one go. However one thing we may need to keep an eye on in the future is memory space as this object will only get cleasred down when a new report is running and so we may have some memory clearing up and disposal jobs to do in the future.