JIRA-4057,4058, 4059 - Cross Tabs not displaying correct data.
Had to pass in the columns array so that it knows at what column to start using as data in the series and which to use as the category. Cross tabs are a natural thing for charting however custom reports are not, so the logic used is that it finds the left most numeric column and assumes that all columns after that are also numeric. In the real world you'd only ever output a summary report to a chart, but if you do want to chart out (for example applicant report) then you will get garbage data out.

Formatting of the output is a pain in the proverbial as we can't just pick up on Excel styles (aspose would seem to have been written for Excel 97). The colours chosen do not look as nice as the native ones in Excel, but we're stuck with aspose for the moment)