JIRA-4124 - The xsi info was appearing on each node rather than at the root. This was caused by the late addition of the XSINIL attribute to handle nil values, but had ommited the root attribute.
Also as part of this I've made the root node name customisable through the custom header attribute in the definition. Note: As this will default to 'custom' when you select xml for the first it may appear disabled on existing definitions from previous builds (select another output type and back to xml to remedy). All new xml output defs will set this value correctly, so a custom should never experience this quirk.