JIRA-4141 - The last run date does not look at successful run.
The fix for this does not involve changing that function (that will still return the last run date regardless of successfullness or not). Instead the export defintion has now inclued a checkbox which adds all of the columns in the export to append to the selected filter and checks those against the audit log - if any of those columns have been changed then the record will form part of the export.

If you have unaudit columns in the list (which is why the grid view has changed to display the audit status) then they wil never trigger that record to be in the export - if the user wants to use this then they really need to configure the audit columns in system manager.