JIRA-4184 - Wrong details being picked up on mobile devices. Was caused by having the reading of the cookie in the new evbent of the loginviewmodel which meant everytime it was referenced it would reload. Swapped to call readcookie manually. To make things a bit more confusing the windows authent was removed for mobile devices, but was actually being populated with the LOGON_USER session variable and hence forcing a window login - namley the logon details that the website prompted for was being used.
Now word of warning here. Safari is rubbish. It will remember the authentication details for the website and NEVER give you the option to forget - yes other sites, like Facebook and so and so can be reset, but anything where IIS has prompted you are remembered forever. Google it! (I bet this was a directive direct form Steve Jobs himself just to piss off windows users.

Use Windows authentication now always enabled regardless of browser or device. The setting is calculated from the IIS setup (anonomous off, asp.net impersonate on, basic or windows authent on)