JIRA-4199, 4200 - No error message being displayed if user does not have permissions.
It was setting the correct message, but the norecords flag was not being set and so the util_run didn't know anything about it.

As part of this and it briefly mentioned in the jira, I've tidied up calendar repots as this had its own (and somewhat flawed) what of dealing with errors. (in simple words I removed this chunk of code)

Also moved the displaycolumns initialisation as it was causing problems when sorting out the display properties on a report grid if there were no columns (i.e. report didn't run successfully)