JIRA-3106, 4180, 3106, 4017 - Wierdness with regional dates settings and different browsers.
Dates should now display is their correct regional settings for reports, prompted values, record edit, and event log. Report definitions and filters calculations have not been altered as there's still a ton of ActiveX lying around in these forms (these should also have the date picker implemented at some point)

Where the regional settings come from is browser dependant, also how they 
interpret those is also hit and miss. Also different controls behave differently. 
For example - the datepicker is formated by calling  setDatePickerLangaue which then uses its own internal set of region data. Validation of dates takes the format from culture setting which we need to return from the server during login (typically dd/MM/yyyy, MM/dd/yy (or yyyy-MM-dd if you're Swedish). The tableTogrid function uses d/m/y - all lowercase or d/m/Y if you want a four didgit year. Two digit years! Its 2014 and still we come across Millenium Bug issues!

As some broswers do not expose their locale date settings we have to rely on the getculture from the server, which is why we populate the txtLocaleCulture, pass this upto the server and then  render the localedateformat when we paint the main form. This way we know that the server and client are talking the same lingo. It will cause problems if a user changes regional settings halfway through a session and don't logout and back in again (unlikely and I'm not going to losse sleep over this.)

There are some areas that have been "fixed" with datepickers (test in filters and filtered add in picklists) but have only been tested in IE as there are still 
ActiveX on these screens. (I think there are also still outstanding issues with 
filtered add (seperate fault)

There are a number of || to get the userlocale and dateformats - this is because some of these functions are called by forms that appear in separate windows. I don't really want to go through changing the logic if in subseqent versions which ditch these windows and replace with divs.

There is a blur event on the datepickers which che3cks the validity of a date and warns the user if necessary. It does not block the save however, so if the user ignores this and saves they will get a "Conversion failed when converting date and/or time from character string." message back from SQL. Ideally we should have some nice pre-validation with some popup "toasters", but timeframes do not allow.