JIRA-4214
Yuck, Super yuck. Horrid. Yuck, yuck, yuck. This was basically to do with JIRA-4021 whereby I re-imagined the transactiontrigger table because it was causing deadlocks during multi-user. This was all fine and dandy however there were a few "special" functions (absence between two dates in this case) that are generated in the system manager and passed through to the system framework which tacks them into the main trigger. These were still pointing at the old style table and hence the trigger sequence was going a bit Pete Tong. Remedy - just swap them over. (I say just - three f*in hours to trace when its sunny outside and I could have enjoyed a beer garden)

This was a horrid one to trace as the trigger that was starting off in absence (insert) which calls the absence (update), which was calling the personnel (update), which scrolls through the child tables (updates) until it hit working patterns (update) which had these special functions - which were then calling absence (update), which was clearing the #instabletrigger and hence the validation was getting bypassed as it should only be called on the very first table that kicked the whole process off. Basically func8.0 is to database normalisation what Hannibal Lector is to vegetarianism.

Also of note (and I'm sure this would yield some performance gains if we were to implement) is that when a trigger fires off an update statement it calls the underlying trigger even if no records match the criteria. Caused much me to scratch my head when UPDATE blah WHERE 1=2 actually fired off some code! No actual records are affected but the trigger and sps will get recompiled and will mess up the query optimisation.

Not all records would cause the error - if there were existing absence records then that was sufficient to limit the range of the update statement - see comment above.

Anyway off to catch those last rays of sunshine with a nice cool beer.