Added some logic to enable/disable the relevant options when selecting all/picklist or filtered records.
Because of the way razor binding works we cannot use the disabled property as there is "clever" code in the background that stops disabled stuff being added to the bound object, so instead I'm using the readonly atribute and setting some styling to mimic the effect of a disable.

There's the beginning of some duplicate code at the util_def level. This in subsequent changesets needs tidying up to either use a layout page or chuck into a bundle or something else.

Also added validation to the related tables parent record selection.

Added a EllipseButton helper as setting the disabled property inline was messy. (Also allows us to make changes in a more consistent manner in the future. (will move more code over to this style in subsequent changesets)

Fixed up bug with filter and picklist names for parent names being the wrong way round (reportRepository)

Modified the updatescript to include a change omitted in previous changeset.

Note: - You will need to run the latest script on any db you wish to view these changes on.