You can now select and save columns for mail merges and custom reports. Add All, Move/Updown remove still to be done.
Baiscally it works by having the two grids (one for availabel, one for selected) and moves rows between the two as the user desires.

The extra details (heading, istotal etc.) are stored in the grid (and this is posted back to the server), but there are also place holders on the view which are populated from the grid when the user selects a row. There is also a onblur which then updates the grid if they alter any of these values.

Yup, I think this can be code reviwed, so long as your aware of the above limitations.