Misc issues on the sort order tab 1) Was displaying all items for all reports (only custom reports needs everything) 2) Name in grid not being updated when new/add of sort order 3) POC, BOC, VOC and suppress were not updatable (and also not being set correctly when drilling down into detail. And the repository was setting the wrong fields. 4) Asc/Desc was displaying code instead of nice text. Use of formater property 5) Drilling down into sort order column was not selecting current column (was defaulting to top) 6) There was no check for the order having already been added. 7) Mail merge retrieval was not picking the correct column ids (stored proc error - reading wrong column)
The move up and move down is a bit flay at the moment because of the way the click/drag functionality is fighting against the programatic way of doing it. I'm sure there's a simple fix for thius, but I can't see the wood for the trees at the moment. This will also affect the move up and move down in the column selection screen.