Added jQuery, jQueryUI, jQueryValidation, jqGrid and Modernizr via NuGet. Even though there was a refrence for nuGet to latch onto they were not managed via nuGet and so there would be conflict when NuGet tried to manage the solution. A consequence of this is that we are foced to use the folder location that NuGet decides (apparently they are going to fix in this future versions, but for the moment we are stuck).
There were some quirks with the slightly newer version of jQuery.
1) .live has been deprecated. Replace with .on (menu.js)
2) $(".accordion").accordion("resize")  has been deprecated - replaced with refresh.
3) $.browser.msie has neen deprecated in jQuery 1.9.0. This was refrerenced in the officebar.js, however as this is a non supported 3rd party and there was no update, so I've manually tweaked this. I've renamed the file and put some comments at the top to reflect this.

There is likely to be further work associated with this as we have a number of .js files which would appear to be redundant/out of date and perhaps also be controlled via NuGet. Also there is probably a lot of clutter in the content folder.
Also note that at this stage I have not brought jQuery and jQueryUI up to the very latest until I've had a proper opportunity to test that there aren't any conflicts.