1. Removed some unnecessary divs and fieldsets 2. Report options in custom reports not retrieving from the database 3. Report "to discuss" from the options label 4. Implemented the missing functionality for defintion of calendar reports (Description3, GroupBy Description, Region 5. Converted Description1 and Description2 for calendar reports to use columndropdown helper 6. Changed the column dropdown helper to actuall filter by the specified datatype rather than just pretending. 7. Changing the base table for calendar reports not picking up on whether there are events or sort orders specified and also not clearing them 8. Selecting an event filter on calendar event details was trying to use old style popup and crashing. Converted to new generic dialog popup
Notes.
Ideally the extra bits of defintions code for calendar reports and custom reports cshould all be handled in _Definition, however this would involve remodelling the model slightly, and I'm running out of time. So this will have to do for the moment. To revisit?

Some love and attention could be given to the styling on this screen