Task 9147: Replace the ActiveX grid on Picklist Definition with a jQuery grid
Final (?) fix to Piclkist Definition; the reported issue was:
After removing a row, the next row is not selected, but the 'Remove' button is still enabled, causing bad things when it is clicked. Either another row should be selected, or the button should disable. Also check what happens when the last row is removed, or all rows are removed.