Added in checks for hidden filters, picklists, descriptions, events and child tables.
Called this functionality ViewAccess instead of a mix of hidden and access and numerous other naming conventions and pass the hidden status around as 'HD', 'RW', or 'RO'. Modified stored procs and models to accomodate this naming convention.

We now pass the filter/picklist type around in uppercase to be consistent - this involved changing the absence breakdown/bradford screen.

Renamed the report saving and validation stored procedures so that they get rid of the _ and also have the same naming convention