Removed the need for the coalicence activeX - instead duiplicated this as a single clsLicence in each of the projects.
A subsequent check-in will actually implment the headcount, and expiry dates. Likewise another check-in will complete the checks in the web module.

I experimented (at serious damage to my sanity) about adjusting the licence key to be base 64 (i.e. allow upper and lower case lettering so as to increase the capacity - however I was getting problems as vb6 doesn't like bit-shifting this much.

I've renamed some of the functions to try and make it clearer how the encyption is working- namely it takes a long value converts it to a base-32 number, these are then stuck together as a string and jumbled (not randomly - despite our best intentions to make it look random)

The licence key upgrade code is written in .NET and held in the system framework. The way it stores empty dates is different to VB6, which is why its got a check for < 1900 rather than an empty check.

This new key structure is already maxed out (hence losing the version key at the start). I've tried to squeeze a bit more out of the dates by converting to a julian(ish) date format (year and day of year) for the expiry date, but there;s only so much you can do with 30 characters)

 I would like at some point to completely revisit how we license our modules, as holding data in a single key like this is difficult. We could look at registering something over the web, keys for individual modules or something else.

In order to test this you will need the latest of all the fat clients and a new system framework. You will also need the latest licence generator.