Added signalR to the project to track concurrency within database. There is now a connection string held within the IIS config which will be used as the default for the database and server for login details. The override on the url is ignored. (and will be tidied up in a subsequent check-in).
This check-in is for the concurrency checks only, and while the base is in place for all the headcount checking the stored proc is not yet created (subsequent task)

SignalR hubs have been added for the system manager messaging, and licence messaging. 

ALERT - You will need to have run the 8.1 update script for the new licence key to work and also have a connection string in the web.config defined for the database of your choice.