Edit and remove buttons were not picking up on whether a row was selected. Now defaults to top row on load of form.
Calcaulations were available in the list of sort orders available. We should only allow calcs or else it breaks the running of the report (the load was zapping them anyway)

The add button was available even if all the availble columns were already added to the report, this was causing display issues if they then clicked on it. Fixed by adding a hidden count field. NOTE: The remove column from report does not as yet decrease this count properly. There are some quirks with remove column that will be addressed in a subsequent check-in.