Remove columns and remove all columns functionality fixed. It now enables and disables the buttons depending on whether available ro selected columns are selected and available.
Previously the removeall looped through each selected column removed and posted back individually. This is now done in one hit and reloads the grid (this causes a small loading time delay, but is still quicker than before)

Converetd colections to a List so that the ".RemoveAll" functionality can be used.

Tidies up the sort orders removal so that these buttons also enable and disable correctly.

NOTE: There are still some quirks with the remove/add all - which will be fixed in a subaeqent check-in (details in related work item)