Removing a column now also removes it from the sort order grid (prompts beforehand)
Remove all checks for sort order columns and removes these too (prompts first)

Removing a column now reloaded the available rather than try to insert the selected row back in. First because this sort out a bug with it insertying at the bottom and not being alphabetical, and second because it was getting confused becasue when its in the selected list it has the table name appended.