Was using the wrong index when adding new sort orders so that when you removed it it could find it.
The server side columns were not being cleared out when a child table was removed meaning they were still available to select from a sort order selected.

The removal of tables was not removing not refreshing the avialble tables list. So I've rejigged the requestRemoveAllChildTables function  to avoid my old error of jquery getting causght in a recursive loop.