The child table button was enabled even if there were no available tables (i.e. a lookup or a child). It worked the first time you went into a definition, but broke if you changed the base table.
While I was testing teh change base table I noticed it was broke for mail merges because of the reseting of the parent field data. So I worted this too.

I've modified it so that data is now returned from the controller action during a post and passes this through to the follow on function specified. This might be auseful technique for fiddly other stuff in the future.