Access Grid is greyed out even for owner if the same session has been logged into by a non-admin/owner of the report. (and the report viewed by that non-owner)
The controls are editable even if you have only read only access.

One of the quirky problems here is that objects generated through the helpers don't seem to be visible when the function() initially loads. Workaround to this is to call them via their class/object type instead of just via their base form.

I've left the double click event active for grid so that the drill down details can be viewed (although these should be read only)