There is a further requirement to add an xsd element within the exported file. At this time I have addressed the requirement for both the template transform file and the xsd to allow the ability to strip off the path.
Extra bugs fixed. (Just happened to notice!)
1. Put an isnull check around the output as it crashed if you had audited changes only and there were no records.
2. Clicking the clear template did not enable the ok button.

Botch alert. I'm using the xmlnamespace to generate the extra bit of data I'm after, however sql by default adds this as xmlns:noNamespaceSchemaLocation= whereas our requirements are xsi:noNamespaceSchemaLocation=, so I've botched the Data Manager to swap this text around.

As there are new columns in export definition table you will need to run the latest update script.

FYI - This will also be retrofitted to a hotifixed 8.0 version of the code and so the bit of the 8.1 update script needs applying to any 8.0 database. This has been created and sits with the hotfiexed 8.0 exe, but is not attached in this changeset.