Added functionality that allows columns to be multi-selected when added or removed from the column grid in report/mail merge definition.
Rejigged the post remove columns to submit the whole lot in one go as there was a lag in getting the available columns to refresh. Created a new ReportColumnCollection class to implement this.

There's a before update to allow for a shift select to work. This also needed to accomate the code to update the column properties grid.

Wrapped the column properties in it own very special div so that I can disable the whole thing in one hit.

When I come to removing items from the available or selected grids I work backwards. The reason for this is that the declared selectedRows is actually a byRef rather than a ByVal and so when you issue a delRowData it was automatically redimensioning the selectedRow array. Nice! Google hints that this may be a bug - but its been there a while and I think it's likely to stay. Traversing the array from the bottom to the top gets round this problem.

Renamed the RemoveSelectedColumn functions to be plural as this now better fits their functionality.

In case you ask the  $("#SelectedColumns").jqGrid('hideCol', 'cb'); line is to hide the checkboxes that typically are visible when you activate the multiselct on a jqgrid.