Group with next now enables for all columns except the very bottom column in the list.
Also group with next and Hidden are mutually exclusive. If either of these are checked then average, count and total are disabled. Care has to be taken so that when they are undisabled they still oibserve the rule that they are only enabled for numeric columns which is why I've split the select row function out into smaller bits.