The defsel was losing focus when saving a new definition or copying an existing definition.
Its still a bit yucky because in my humble opinion if you click on something in the accordion it always jumps to the top item in the list. This is however consistent with expresssions, picklists et al. Ideally we should look at revisting this once the other items are in mvc, we can then have a nice navigation class rather than relying on session variables.