Moved the ignore aggregate from definition to the columns tab. I agree that it does make sense. I have not moved the summary, although I can see an arguement for that moving too.
I've also taken the opportunity to widen the selected columns list. This should not be considered styling for this screen - merely an emergency fix as half the column name was permanently hidden.

The tfs issue says "blowing up" on save if aggregate was set. I was unable to replicate this (I suspect it might be a work in progress issue with the build QA had). Unlike Data Manager the validation on this will cause a save failure message and not an immediate popup warning. This is by design as its felt this given a better consistent user story.