Repeated removing and adding of child tables and calendar event entries was getting the ids mixed up and hence the appearance of duplicates events and child tables.
Code fix is relatively simple. Instead of using a count to allocate the next id, scan the collection for the maximum value and apply one to that. Made a bit tricker by the calendar report one as I've had to convert the EV_xxx index to a number and then back to a string at database saving time so that it doesn't screw up the definition in the Data Manager.