No validation for empty event title and no validation for empty end date or duration column if end date or duration selected.
Notes:
Its slightly different to the way of validating the main definiton views as that does a post to the server and uses the data annotations on the model, however we are not posting at this stage hence its just a bit of hardcoded validation.

This changeset does not look at validation of lookup columns if there are none available - this is a slightly different issue.

This could be reviewed again if/when we implement client side validation, some nice highlighting of errors next to the control and/or automatic disabling of the ok button might be desirable (not doing at this stage though)