Changed the sort order grid to only display column names for calendar reports. Since you can only have one table this is making better use of the space.
Track use of the amount of sort orders avaialble. The logic here is different for mail merges/custom reports and calendar reports. Calendar is all columns for the base table, the others pick from the selected columns.

When editing orders it did not track whether it was a new order and it would theerfore not reselect the correct row in the grid if they went in and changed the column that the sort order was based on.

Made the sort orders property readonly and calculated based on teh contents of the sort order.

Refreshed teh amount of available sort orders when changing the base table.

These changes roll across all report types.