Changed logic around a bit so that columns and tables are only populated for lookups if lookup types are selected.
Because of fun and games disabling controls and that different broswers handle hide of select option items differently I've put some dummy lookups which I toggle on and offf to pretend that we have disabled items.

The GetAvailableCharacterLookupsForTable should really be combidned with GetColumnsForTable, but I don't want to risk accidental changes in the project at this stage that would affect other areas.

Added an ability to put a follow on function for submitForm. I'm using it in this instance to do some refreshing after the calendar event base table is changed.

I've made some changes in the html part of CalendarEventDetail. If/when it's de-tabled please be careful not to mess up any of the class names on the dummy combos.

Renmaed some of the functions in CalendarEventDetail to reflect if they are change functions or simply refresh.

Changed the LengendTypeName that's returned to the main grid to differenciate between a lookuptable text and a simple charcater key.

There were two $(function () { on the CalendarEventDetail - tidied this up so now its only one.