Renamed the enums as the concurrency check only applies to the DMI and DMI single modules
In order for the headcount to work we are reliant on the overnight job task to be implemented (tfs-11125) which at time of check-in is not completed. For the moment a figure in asrsyssystemsettings has been engineered.

A licence key to test the headcount is "6NGFUG-ULUAP7-MAJFH7-3AAA87-AANF6J"

The warning should only appear once a week for a headcount nearing threshold. (11226). This has also yet to be completed.

I've used the banner/sscx for the hub controller as this is something that exists for both ssi and dmi modules.