Added a new table and stored proc to track the warning of users. In a future check-in the Fat clients will also look at these functions.
Two types of warning are allowed at present, but I've included the type as an enum for future expansion. The rate at which the warnings are displayed to the user is passed in as a parameter (7 days for headcount, 1 day for licence expiry (the spec says every Monday for headcount, but I'm going to negotiate that one later)

You will need ro run the sysMan update script for this to work. (npg_openhr8 is done)