Now forces the user to enter a licence key as part of a system manager save.
The licence key has been amended to store the headcount and SSI in the same blob of data as they are mutually exclusive. Also included is a optional validation for the creation date (a key is now only valid for one day only (or if you want to be picky that day of the year, so it'll work for the same time next year)

The security manager licence has been modified so that the user simply enters a licence key and the modules are calculated automatically.

This has required so extra characters in the key and also now included is a checksum to try and prevent tampering (which was rather easy before)

A check is made upon login for a valid licence key. Anything invalid and the user is altered.