Deleting a nine box was erroring. The problem was that the check for utilityType = 35 was not rolled into all relevant stored procs.
As there were further checks in spASRIntGetUtilityBaseTable and  spASRIntGetUtilityName I took the liberty of removing them procs once and for all. (A small tweak for getname as it was called from get def properties - so that was tweaked too)

Also there were some references to "NINEBOX" and "NINEBOXGRID" these are now cleaned up.