There was a stored procedure that's created to track the headcount component of the licence - unfortunately the code for that was in System Manager and the licence change code in the security manager, so changing the headcount values or the licence type was not being factored in the stored proc (unless we forced them to go into System Manager after changing the licence key.
The remedy for this is to move the licence from Security to System Mgr. Fairly straightforward for the most part, just lift and drop frmLicence and frmLicenceKey from one project to another.

There were of course a few tweaks required - the global variables for tracking which modules are licenced needed to be more dynamic and not just loaded at login (actually I've removed these globals and just reference IsModuleEnabled where appropriate.

The security manager no longer allows any kind of access if the licence has expired. Conversely the system manager does now allow access, but only the licence menu item is enabled.

The headcount stored proc is now encrypted (is not going to stop anyone who really wants to bypass the licence checks, but it might at least make them stop and think.)

In the modules menu all items are now always visible, but enabled or disabled based on licence settings.

The calcvulate headcount proc is regenerated and executed after a licence key change