I could never make this happen in my environment, but occasioanlly some clients were getting "Sequence contains no matching element" when logging in. I suspect this may be just a delay in hooking up to licence hub, so now I've disabled the login button and only activate once we have a successfully connected to the hub (This process sends a message down to the client to actiavte the login button)
DISCLAIMER - Risk. If my suspicions are wrong and something else is stopping the signalR client from initialising properly then the login button will remain permanently disabled. Since (apparently) a couple of the other devs have seen this happen on their machine previously hopefully we can prove that I am right before we ship to QA (I'm likely to get my head kicked in if we ship a version with the login button disabled)