Sorry, but this one is going to be a bit like War and Peace.
The 'Use as workflow target identifier' checkbox should now appear for stored data elements and the properties drilldown for record selectors. There's a lot of quite hairy logic going round here with regard to the resize and disaplay logic depending on the element type. Basically we look for the existence of WFITEMPROP_USEASTARGETIDENTIFIER and then decide whether we display the checkbox and the item in the properties grid.

I have not implemented any extra validation for the selection of these two new checkboxes in this changeset. The spec is a bit fluffy at the moment, so this will be addressed in another changeset/task.

There was a reliance on the external ActiveX controls (stored data element and dummy grid control. In order to avoid all sorts of fun and games with DLL hell I've moved these two controls directly into the System Manager vbp. Part of this involved removing some duplicate public enums. These controls were then deleted off the designer page and the new internal copy re-added. I manually made sure each of the settings were identical to the original (the .index and .visible properties are the only ones not the default) Also I've now deprecated these controls and put them into the obsolete folder in Source control

The horrible way that the values are populated means expanding the 80 odd dimension properties array and adding a 79th enum for property type. I am so, so sorry about this. I wish I could have the time to rip all this crap out and replace with a nice strongly typed collection.

When coparing the changies (mainly the .frm VB seems to be having some fun and games with changing the case of certain properties. I don't know why its doing, this, I think its just some disagreement with VB6 and TFS. VB6 is not case sensitive, so this will not cause a problem.Ditoo for controls at the top it often changes the file path (..\program files) to \dev3rd Party etc), as VB6 actually uses class ID, this should also not be problematic so long as the cls ids are the same.

It goes without saying (although I am saying it) you will need to run the included update script in this changeset